requirements.txt
pyannote.audio==3.1
pydub==0.25.1